move methods to Notices class
